+++ projects/ontimize-web-ngx/src/lib/components/input/phone-input/o-phone-input.component.spec.ts
@@ -7,6 +7,7 @@ import { TranslateService } from '@ngx-translate/core';
 
 import { APP_CONFIG, AppConfig } from '../../../config/app-config';
 import { appConfigFactory, OPermissionsModule } from '../../../services';
+import { TestUtils } from '../test/test-utils';
 import { InputTestUtil } from './../test/input-test-utils';
 import { OPhoneInputComponent } from './o-phone-input.component';
 import { OPhoneInputModule } from './o-phone-input.module';
@@ -29,7 +30,7 @@ describe('OPhoneInputComponent', () => {
           useClass: TranslateService,
           deps: [Injector]
         },
-        { provide: APP_CONFIG, useValue: InputTestUtil.mockConfiguration() },
+        { provide: APP_CONFIG, useValue: TestUtils.mockConfiguration() },
         { provide: AppConfig, useFactory: appConfigFactory, deps: [Injector] }
         // ...INTERNAL_ONTIMIZE_MODULES
       ]